Move BytesRefUtils into testing
It is only used in testing code.
